requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.25
requests==2.31.0
typer==0.9.0
streamlit==1.28.1
kubernetes==28.1.0
python-dotenv==1.0.0
pandas>=1.5.0,<3.0.0
httpx>=0.24.0,<0.28.0